+++ packages/lit/lit-jsx/src/compiler/transform-jsx.ts
@@ -1,18 +1,19 @@
-import { type PluginPass } from '@babel/core';
+import type { PluginPass } from '@babel/core';
 import type { NodePath, VisitNode } from '@babel/traverse';
-import * as t from '@babel/types';
+import type { JSXElement, JSXFragment } from '@babel/types';
+import { isJSXElement, isJSXFragment } from '@babel/types';
 
 import { getTemplateType, isJSXElementStatic, isJSXFunctionElementComponent } from './compiler-utils.js';
 import { CompiledTranspiler, TemplateTranspiler } from './transpiler.js';
 
 
 export const transformJSXElement: VisitNode<
-	PluginPass, t.JSXElement | t.JSXFragment
+	PluginPass, JSXElement | JSXFragment
 > = (path): void => {
 	// If the parent is a JSX element, we do not need to transform it.
 	// The below condition will handle the case where the JSX element
 	// is nested inside another JSX element.
-	if (t.isJSXElement(path.parent) || t.isJSXFragment(path.parent))
+	if (isJSXElement(path.parent) || isJSXFragment(path.parent))
 		return;
 
 	// If the parent is not a JSX element,
@@ -21,38 +22,20 @@ export const transformJSXElement: VisitNode<
 };
 
 
-const processJSXElement = (path: NodePath<t.JSXElement | t.JSXFragment>) => {
-	//console.time('Analyzing if JSX Element is static');
-	const isStatic = isJSXElementStatic(path);
-	//console.timeEnd('Analyzing if JSX Element is static');
-
-	//console.time('Analyzing template type');
-	const templateType = getTemplateType(path);
-	//console.timeEnd('Analyzing template type');
-
-	//console.time('Analyzing if JSX Element is a function component');
+const processJSXElement = (path: NodePath<JSXElement | JSXFragment>) => {
 	const isFunctionComponent = isJSXFunctionElementComponent(path);
-	//console.timeEnd('Analyzing if JSX Element is a function component');
 
-	if (isFunctionComponent) {
-		//console.time('Functional Component Transpilation');
-		const cmp = new TemplateTranspiler().createFunctionalComponent(path);
-		//console.timeEnd('Functional Component Transpilation');
+	// A root function component needs to be immediately created.
+	if (isFunctionComponent)
+		return new TemplateTranspiler().createFunctionalComponent(path);
 
-		return cmp;
-	}
+	const isStatic = isJSXElementStatic(path);
+	const templateType = getTemplateType(path);
 
-	if (isStatic || templateType !== 'html') {
-		//console.time('Static Transpilation');
-		const cmp = new TemplateTranspiler().start(path);
-		//console.timeEnd('Static Transpilation');
+	// If the element is static, we need to use the template transpiler.
+	if (isStatic || templateType !== 'html')
+		return new TemplateTranspiler().start(path);
 
-		return cmp;
-	}
-
-	//console.time('Compiled Transpilation');
-	const cmp = new CompiledTranspiler().start(path);
-	//console.timeEnd('Compiled Transpilation');
-
-	return cmp;
+	// A non static, non function component can be transpiled using the compiled transpiler.
+	return new CompiledTranspiler().start(path);
 };